moved exchange components into exchange_components directory and exchanges into exchanges directory
