+++ app/components/Exchanges/Changelly.js
@@ -1,23 +1,23 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import Claim from "./Claim.js";
+import Claim from "../Claim.js";
 import MdSync from "react-icons/lib/md/sync";
 import QRCode from "qrcode.react";
-import { initiateGetBalance, intervals } from "../components/NetworkSwitch";
-import Exchange_Unavailable from "../components/Exchange_Unavailable";
-import { resetPrice } from "../modules/wallet";
-import { sendEvent, clearTransactionEvent } from "../modules/transactions";
+import { initiateGetBalance, intervals } from "../NetworkSwitch";
+import Exchange_Unavailable from "../exchange_components/Exchange_Unavailable";
+import { resetPrice } from "../../modules/wallet";
+import { sendEvent, clearTransactionEvent } from "../../modules/transactions";
 import { clipboard } from "electron";
 import Copy from "react-icons/lib/md/content-copy";
 import ReactTooltip from "react-tooltip";
-import neoLogo from "../img/neo.png";
-import NeoLogo from "./Brand/Neo";
-import BtcLogo from "./Brand/Bitcoin";
+import neoLogo from "../../img/neo.png";
+import NeoLogo from "../Brand/Neo";
+import BtcLogo from "../Brand/Bitcoin";
 
 import { Link } from "react-router";
 import crypto from "crypto";
 import axios from "axios";
-import Changelly from "../modules/changelly";
+import Changelly from "../../modules/changelly";
 import { error } from "util";
 
 // force sync with balance data
