+++ app/routes.js
@@ -4,8 +4,8 @@ import App from "./components/App";
 import LoginNep2 from "./components/LoginNep2";
 import LoginPrivateKey from "./components/LoginPrivateKey";
 import Settings from "./components/Settings";
-import Changelly from "./components/Changelly";
-import ShapeShift from "./components/ShapeShift";
+import Changelly from "./components/exchanges/Changelly";
+import ShapeShift from "./components/exchanges/ShapeShift";
 import SelectExchange from "./components/SelectExchange";
 import ExchangeAddress from "./components/ExchangeAddress";
 import LoginLocalStorage from "./components/LoginLocalStorage";
